TOAST mop-up work: update comments for tuple-size-related symbols such as MaxHeapAttributeNumber.  Increase MaxAttrSize to something more reasonable (given what it's used for, namely checking char(n) declarations, I didn't make it the full 1G that it could theoretically be --- 10Mb seemed a more reasonable number).  Improve calculation of MaxTupleSize.

+++ src/backend/utils/adt/varbit.c
@@ -107,7 +107,7 @@ zpbit_in(char *s, int dummy, int32 atttypmod)
 
 	if (len > MaxAttrSize)
 		elog(ERROR, "zpbit_in: length of bit() must be less than %ld",
-			 (MaxAttrSize - VARHDRSZ - VARBITHDRSZ) * BITSPERBYTE);
+			 (long) ((MaxAttrSize - VARHDRSZ - VARBITHDRSZ) * BITSPERBYTE));
 
 	result = (bits8 *) palloc(len);
 	/* set to 0 so that *r is always initialised and strin is zero-padded */
@@ -338,7 +338,7 @@ varbit_in(char *s, int dummy, int32 atttypmod)
 
 	if (len > MaxAttrSize)
 		elog(ERROR, "varbit_in: length of bit() must be less than %ld",
-			 (MaxAttrSize - VARHDRSZ - VARBITHDRSZ) * BITSPERBYTE);
+			 (long) ((MaxAttrSize - VARHDRSZ - VARBITHDRSZ) * BITSPERBYTE));
 
 	result = (bits8 *) palloc(len);
 	/* set to 0 so that *r is always initialised and strin is zero-padded */
